Refactor: extract tools into registerTools(server, getAgent)
Move the 21 server.tool() registrations out of src/index.ts into a new
src/tools.ts module, exported as registerTools(server, getAgent). The
getAgent callback is evaluated per tool call, which lets future transports
(e.g. an HTTP/OAuth remote server) provide a per-session AtpAgent instead
of the current module-level singleton.

Also moves mcpLog / mcpErrorResponse / mcpSuccessResponse helpers from
index.ts to utils.ts so tools.ts can import them without a cycle. They
remain re-exported from index.ts for backwards compatibility.

No behavior change for the stdio server: index.ts still owns the single
AtpAgent and passes () => agent as the provider, so tools observe
identical semantics. Verified by tools/list returning all 21 tool names.
